+++ CHANGELOG.md
@@ -11,6 +11,14 @@
   * `getCurrentDeviceId`
   * `getDeviceIDType`
   * `changeDeviceId`
+* Mitigated issues where:
+  * session was ending with device ID change without merge, when consent was not required and manual session control was enabled on Android.
+  * session was not starting after device ID change without merge, when consent was not required and automatic sessions were enabled on Android.
+  * consent information was not sent when no consent was given during initialization on iOS.
+  * session could have started if the SDK was initialized on the background and automatic session tracking was enabled on iOS.
+  * session did not end when session consent was removed on iOS.
+  * disabling location did not work on iOS.
+  * orientation info was not sent during initialization on iOS.
 
 * Updated underlying Android SDK version to 24.7.1
 * Updated underlying iOS SDK version to 24.7.1
@@ -28,6 +36,14 @@
   * `getCurrentDeviceId`
   * `getDeviceIDType`
   * `changeDeviceId`
+* Mitigated issues where:
+  * session was ending with device ID change without merge, when consent was not required and manual session control was enabled on Android.
+  * session was not starting after device ID change without merge, when consent was not required and automatic sessions were enabled on Android.
+  * consent information was not sent when no consent was given during initialization on iOS.
+  * session could have started if the SDK was initialized on the background and automatic session tracking was enabled on iOS.
+  * session did not end when session consent was removed on iOS.
+  * disabling location did not work on iOS.
+  * orientation info was not sent during initialization on iOS.
 
 * Updated underlying Android SDK version to 24.7.1
 * Updated underlying iOS SDK version to 24.7.1